refactor: tweak component names and directory structure
