+++ src/components/App.js
@@ -30,7 +30,7 @@ import classes from './App.module.css'
 import { default as DetailsPanel } from './DetailsPanel/DetailsPanel'
 import { default as DialogManager } from './Dialogs/DialogManager'
 import DndContext from './DndContext'
-import { InterpretationsModal } from './Interpretations/InterpretationsModal'
+import { InterpretationModal } from './Interpretations/InterpretationModal'
 import Layout from './Layout/Layout'
 import LoadingMask from './LoadingMask/LoadingMask'
 import { default as TitleBar } from './TitleBar/TitleBar'
@@ -260,7 +260,7 @@ const App = ({
                                             />
                                         )}
                                         {current && (
-                                            <InterpretationsModal
+                                            <InterpretationModal
                                                 visualization={current}
                                                 onResponseReceived={
                                                     onResponseReceived
