+++ src/components/Interpretations/InterpretationModal/Comment.js
@@ -2,11 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import { IconEdit16 } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
-import {
-    Message,
-    MessageIconButton,
-    MessageStatsBar,
-} from '../../common/index.js'
+import { Message, MessageIconButton, MessageStatsBar } from '../common/index.js'
 import { CommentDeleteButton } from './CommentDeleteButton.js'
 import { CommentUpdateForm } from './CommentUpdateForm'
