+++ src/components/Interpretations/InterpretationModal/CommentDeleteButton.js
@@ -3,7 +3,7 @@ import i18n from '@dhis2/d2-i18n'
 import { IconDelete16 } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { MessageIconButton } from '../../common/index.js'
+import { MessageIconButton } from '../common/index.js'
 
 const mutation = {
     resource: 'interpretations',
