+++ src/components/Interpretations/InterpretationModal/index.js
@@ -1,0 +1,1 @@
+export { InterpretationModal } from './InterpretationModal.js'
