+++ src/components/Interpretations/InterpretationModal/styles/InterpretationModal.module.css
