+++ src/components/Interpretations/InterpretationModal/styles/InterpretationModalContent.module.css
