--- src/components/Interpretations/InterpretationsModal/InterpretationsThread/index.js
@@ -1,1 +1,0 @@
-export { InterpretationThread } from './InterpretationThread.js'
