--- src/components/Interpretations/InterpretationsModal/index.js
@@ -1,1 +1,0 @@
-export { InterpretationsModal } from './InterpretationsModal.js'
